requirements.txt
# Django
pandas
scipy
numpy
matplotlib

# Django Apps
django-csvimport
djangorestframework

# Algos
statsmodels
tad
